CR-1129630 Move query requests exceptions into query.h. Remove ready flag changes from mgmt driver. Update catch conditions for query default functions.
